Allow the formatting Verb to be given.
This allows the user to specify which float format they want to use. It will
panic if an invalid Verb is given.

Also Out-Of-Bound values can be formatted specially. This allows for
especially large or small values to be formatted so as to be visible while
using a more pleasing format for the majority of values.

Also you can now choose to trim trailing zeroes after the decimal place. Note
that the first zero immediately after the decimal place is kept.