Change the NewReport... funcs so that they can't return an error
We do this by making the first Col a parameter so it won't compile without at
least one column. This way we don't need to check for errors and so we don't
need a NewReportOrPanic func.

Also added a new func: StdRpt which just takes the Col arguments and sets the
writer to os.Stdout and uses a minimal header (with no parameters)